requirements.txt
langchain
langchain-google-genai
python-dotenv
pydantic
fastapi
uvicorn
